refactor: split GMBridge into modular files (world_step, markers, inventory, bottle_map, survey_cache, world_scan, encounters, shared) and export shim from gm_bridge/index.js
