+++ package.json
@@ -5,6 +5,8 @@
     "ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa"
   ],
   "scripts": {
+    "lint:strict": "npm --workspace tiny-roguelike run lint:strict",
+    "build": "npm --workspace tiny-roguelike run build",
     "ci": "npm --workspace tiny-roguelike run ci",
     "acceptance:phase6": "npm --workspace tiny-roguelike run acceptance:phase6",
     "acceptance:phase0": "npm --workspace tiny-roguelike run acceptance:phase0",
